// examples/crud-api/generated/task.gen.ts
export const TaskDataSchema = z.object({
  id: z.string(),
  title: z.string(),
  done: z.boolean().optional(),
  createdAt: z.string().datetime(),
  updatedAt: z.string().datetime(),
}).strict()
export type TaskData = z.infer<typeof TaskDataSchema>